Move method to better suitable class #2
